requirements.txt
cstruct==6.1
pyserial==3.5
